refactor: moved services logic to resolvers file
